# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trivelo Hotel Booking API",
        "description": "Comprehensive hotel booking system API with multi-role authentication",
        "contact": {
            "email": "support@trivelo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "List user bookings",
                "description": "Get authenticated user's bookings with filtering and pagination",
                "operationId": "4535e967ef15bc76dadf1efa40192f0d",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by booking status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "confirmed",
                                "pending",
                                "cancelled",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "hotel_id",
                        "in": "query",
                        "description": "Filter by hotel ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "check_in_date",
                                "check_out_date",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookings retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "description": "Create a new room booking for authenticated user",
                "operationId": "43e6dcca842886074084b8db86af7802",
                "requestBody": {
                    "description": "Booking creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBookingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingDetailResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or room not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - Only customers can create bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get booking details",
                "description": "Get detailed information about a specific booking",
                "operationId": "c383a28985575aa1a16f783b2f762554",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Booking ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingDetailResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - Can only view own bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BookingCreateRequest": {
                "title": "Booking Create Request",
                "description": "Create booking request",
                "required": [
                    "room_id",
                    "check_in_date",
                    "check_out_date",
                    "guests_count"
                ],
                "properties": {
                    "room_id": {
                        "description": "Room ID",
                        "type": "integer",
                        "example": 1
                    },
                    "check_in_date": {
                        "description": "Check-in date",
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-01"
                    },
                    "check_out_date": {
                        "description": "Check-out date",
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-05"
                    },
                    "guests_count": {
                        "description": "Number of guests",
                        "type": "integer",
                        "minimum": 1,
                        "example": 2
                    },
                    "special_requests": {
                        "description": "Special requests",
                        "type": "string",
                        "example": "Late check-in requested",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "CreateBookingRequest": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BookingCreateRequest"
                    }
                ]
            },
...
            "BookingListResponse": {
                "title": "Booking List Response",
                "description": "Paginated booking list response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Bookings retrieved successfully"
                    },
                    "data": {
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/BookingSummary"
                                }
                            },
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 2
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 25
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "BookingDetailResponse": {
                "title": "Booking Detail Response",
                "description": "Single booking detail response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Booking retrieved successfully"
                    },
                    "data": {
                        "$ref": "#/components/schemas/BookingDetail"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Error API response format",
                "properties": {
                    "success": {
                        "description": "Operation success status",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    },
                    "errors": {
                        "description": "Detailed error information",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "The email field is required.",
                            "The password must be at least 8 characters."
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}